Silently ignore unrecognized commands
We used to have a 'raise NotImplementedError' here, and as soon as
this was deployed we hit that error, from an issue that contained some
Python warning messages, which start with an absolute path, and the
path starts with a /. Whoops. I don't think there's any better way to
handle these except to just ignore them.